Pipeline ------------ 1. Add migration in pre-deploy 2. Add test stage for dev -------------------

+++ api/api_sources/sources/database/__tests__/db.spec.ts
@@ -77,18 +77,19 @@ describe('Test for db utilities', () => {
         should().exist(sql);
     });
 
-    it('should load and create sql from schema yaml', () => {
-        const schema = new Test2Schema();
-        const sql = schema.createTable();
-        // console.log(sql);
-        should().exist(sql);
-        schema.table.name = 'sample2_table';
-        schema.saveSchema();
-        schema.createMigrationFile(path.resolve(__dirname, '../database-schema/schema-sqls/sample.sql'));
-        fs.unlinkSync(path.resolve(__dirname, '../database-schema/schema-sqls/sample.sql'));
-
-        // schema.saveSchema(path.resolve(__dirname, '../database-schema/schema-files/sample2.schema.yaml'));
-    });
+    if (process.env.ENVIRONMENT === 'local') {
+        it('should load and create sql from schema yaml', () => {
+            const schema = new Test2Schema();
+            const sql = schema.createTable();
+            // console.log(sql);
+            should().exist(sql);
+            schema.table.name = 'sample2_table';
+            schema.saveSchema();
+            schema.createMigrationFile(path.resolve(__dirname, '../database-schema/schema-sqls/sample.sql'));
+            fs.unlinkSync(path.resolve(__dirname, '../database-schema/schema-sqls/sample.sql'));
+            // schema.saveSchema(path.resolve(__dirname, '../database-schema/schema-files/sample2.schema.yaml'));
+        });
+    }
 
     it('should load schema', () => {
         const pathOfSchema = getYAMLFilePath('observation.codes.schema.yaml');
